requirements.txt
streamlit==1.28.0
pandas==2.0.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
plotly==5.15.0
